# docs/swagger.yaml
basePath: /v1
definitions:
...
  handlers.RefreshTokenRequestBody:
    properties:
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
        type: string
    type: object
...
  handlers.Response:
    properties:
      data: {}
      error:
        type: boolean
      message:
        type: string
    type: object
...
  models.Post:
    properties:
      body:
        example: dolore maxime saepe...
        type: string
      id:
        example: 61
        type: integer
      title:
        example: voluptatem doloribus...
        type: string
      userId:
        example: 7
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    name: Artem Kostenko
    url: https://github.com/aerosystems
  description: Simple REST API for CRUD operations with Comments & Posts enities.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: NIX Trainee 5-6-7-8 tasks
  version: "1.0"
paths:
...
  /posts:
    get:
      consumes:
      - application/json
      - text/xml
      parameters:
      - description: should contain Access Token, with the Bearer started
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      - application/xml
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Post'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.Response'
      summary: get all posts
      tags:
      - posts
  /posts/{id}:
    delete:
      consumes:
      - application/json
      - text/xml
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      - description: should contain Access Token, with the Bearer started
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      - application/xml
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Post'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.Response'
      summary: delete post by ID
      tags:
      - posts
    get:
      consumes:
      - application/json
      - text/xml
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      - description: should contain Access Token, with the Bearer started
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      - application/xml
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Post'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.Response'
      summary: get post by ID
      tags:
      - posts
    patch:
      consumes:
      - application/json
      - text/xml
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      - description: should contain Access Token, with the Bearer started
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      - application/xml
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Post'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.Response'
      summary: particular update post by ID
      tags:
      - posts
    post:
      consumes:
      - application/json
      - text/xml
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      - description: should contain Access Token, with the Bearer started
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      - application/xml
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Post'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.Response'
      summary: create post by ID
      tags:
      - posts
  /tokens/refresh:
    post:
      consumes:
      - application/json
      - text/xml
      parameters:
      - description: raw request body, should contain Refresh Token
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/handlers.RefreshTokenRequestBody'
      - description: should contain Access Token, with the Bearer started
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      - application/xml
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.Response'
      summary: refresh pair JWT tokens
      tags:
      - users
...
swagger: "2.0"